+++ traffic_ops/testing/api/v4/deliveryservice_requests_test.go
@@ -23,7 +23,6 @@ import (
 
 	"github.com/apache/trafficcontrol/lib/go-rfc"
 	tc "github.com/apache/trafficcontrol/lib/go-tc"
-	"github.com/apache/trafficcontrol/traffic_ops/testing/api/utils"
 )
 
 const (
@@ -32,6 +31,27 @@ const (
 	dsrRequired  = 2
 	dsrDraft     = 3
 )
+
+// this resets the IDs of things attached to a DS, which needs to be done
+// because the WithObjs flow destroys and recreates those object IDs
+// non-deterministically with each test - BUT, the client method permanently
+// alters the DSR structures by adding these referential IDs. Older clients
+// got away with it by not making 'DeliveryService' a pointer, but to add
+// original/requested fields you need to sometimes allow each to be nil, so
+// this is a problem that needs to be solved at some point.
+// A better solution _might_ be to reload all the test fixtures every time
+// to wipe any and all referential modifications made to any test data, but
+// for now that's overkill.
+func resetDS(ds *tc.DeliveryServiceV4) {
+	if ds == nil {
+		return
+	}
+	ds.CDNID = nil
+	ds.ID = nil
+	ds.ProfileID = nil
+	ds.TenantID = nil
+	ds.TypeID = nil
+}
 
 func TestDeliveryServiceRequests(t *testing.T) {
 	WithObjs(t, []TCObj{CDNs, Types, Parameters, Tenants, DeliveryServiceRequests}, func() {
@@ -56,7 +76,11 @@ func TestDeliveryServiceRequests(t *testing.T) {
 func UpdateTestDeliveryServiceRequestsWithHeaders(t *testing.T, header http.Header) {
 	// Retrieve the DeliveryServiceRequest by name so we can get the id for the Update
 	dsr := testData.DeliveryServiceRequests[dsrGood]
-	resp, _, err := TOSession.GetDeliveryServiceRequestByXMLIDWithHdr(dsr.DeliveryService.XMLID, header)
+	resetDS(dsr.DeliveryService)
+	if dsr.DeliveryService == nil || dsr.DeliveryService.XMLID == nil {
+		t.Fatalf("the %dth DSR in the test data had no DeliveryService - or that DeliveryService had no XMLID", dsrGood)
+	}
+	resp, _, _, err := TOSession.GetDeliveryServiceRequestsByXMLID(*dsr.DeliveryService.XMLID, header)
 	if err != nil {
 		t.Errorf("cannot GET DeliveryServiceRequest by name: %v - %v", dsr.DeliveryService.XMLID, err)
 	}
@@ -64,9 +88,14 @@ func UpdateTestDeliveryServiceRequestsWithHeaders(t *testing.T, header http.Head
 		t.Fatal("Length of GET DeliveryServiceRequest is 0")
 	}
 	respDSR := resp[0]
-	respDSR.DeliveryService.DisplayName = "new display name"
+	if respDSR.ID == nil {
+		t.Fatalf("Got a DSR for XML ID '%s' that had a nil ID", *dsr.DeliveryService.XMLID)
+	}
 
-	_, reqInf, err := TOSession.UpdateDeliveryServiceRequestByIDWithHdr(respDSR.ID, respDSR, header)
+	respDSR.DeliveryService.DisplayName = new(string)
+	*respDSR.DeliveryService.DisplayName = "new display name"
+
+	_, _, reqInf, err := TOSession.UpdateDeliveryServiceRequest(*respDSR.ID, respDSR, header)
 	if err == nil {
 		t.Errorf("Expected error about precondition failed, but got none")
 	}
@@ -77,7 +106,11 @@ func UpdateTestDeliveryServiceRequestsWithHeaders(t *testing.T, header http.Head
 
 func GetTestDeliveryServiceRequestsIMSAfterChange(t *testing.T, header http.Header) {
 	dsr := testData.DeliveryServiceRequests[dsrGood]
-	_, reqInf, err := TOSession.GetDeliveryServiceRequestByXMLIDWithHdr(dsr.DeliveryService.XMLID, header)
+	resetDS(dsr.DeliveryService)
+	if dsr.DeliveryService == nil || dsr.DeliveryService.XMLID == nil {
+		t.Fatalf("the %dth DSR in the test data had no DeliveryService - or that DeliveryService had no XMLID", dsrGood)
+	}
+	_, _, reqInf, err := TOSession.GetDeliveryServiceRequestsByXMLID(*dsr.DeliveryService.XMLID, header)
 	if err != nil {
 		t.Fatalf("Expected no error, but got %v", err.Error())
 	}
@@ -88,7 +121,7 @@ func GetTestDeliveryServiceRequestsIMSAfterChange(t *testing.T, header http.Head
 	currentTime = currentTime.Add(1 * time.Second)
 	timeStr := currentTime.Format(time.RFC1123)
 	header.Set(rfc.IfModifiedSince, timeStr)
-	_, reqInf, err = TOSession.GetDeliveryServiceRequestByXMLIDWithHdr(dsr.DeliveryService.XMLID, header)
+	_, _, reqInf, err = TOSession.GetDeliveryServiceRequestsByXMLID(*dsr.DeliveryService.XMLID, header)
 	if err != nil {
 		t.Fatalf("Expected no error, but got %v", err.Error())
 	}
@@ -101,8 +134,10 @@ func CreateTestDeliveryServiceRequests(t *testing.T) {
 	t.Log("CreateTestDeliveryServiceRequests")
 
 	dsr := testData.DeliveryServiceRequests[dsrGood]
-	respDSR, _, err := TOSession.CreateDeliveryServiceRequest(dsr)
+	resetDS(dsr.DeliveryService)
+	respDSR, alerts, _, err := TOSession.CreateDeliveryServiceRequest(dsr, nil)
 	t.Log("Response: ", respDSR)
+	t.Logf("Alerts from creating a dsr: %+v", alerts)
 	if err != nil {
 		t.Errorf("could not CREATE DeliveryServiceRequests: %v", err)
 	}
@@ -112,7 +147,8 @@ func CreateTestDeliveryServiceRequests(t *testing.T) {
 func TestDeliveryServiceRequestRequired(t *testing.T) {
 	WithObjs(t, []TCObj{CDNs, Types, Parameters, Tenants}, func() {
 		dsr := testData.DeliveryServiceRequests[dsrRequired]
-		_, _, err := TOSession.CreateDeliveryServiceRequest(dsr)
+		resetDS(dsr.DeliveryService)
+		_, _, _, err := TOSession.CreateDeliveryServiceRequest(dsr, nil)
 		if err == nil {
 			t.Error("expected: validation error, actual: nil")
 		}
@@ -127,11 +163,15 @@ func TestDeliveryServiceRequestRules(t *testing.T) {
 		displayName := strings.Repeat("X", 49)
 
 		dsr := testData.DeliveryServiceRequests[dsrGood]
-		dsr.DeliveryService.DisplayName = displayName
-		dsr.DeliveryService.RoutingName = routingName
-		dsr.DeliveryService.XMLID = XMLID
+		resetDS(dsr.DeliveryService)
+		if dsr.DeliveryService == nil {
+			t.Fatalf("the %dth DSR in the test data had no DeliveryService", dsrGood)
+		}
+		dsr.DeliveryService.DisplayName = &displayName
+		dsr.DeliveryService.RoutingName = &routingName
+		dsr.DeliveryService.XMLID = &XMLID
 
-		_, _, err := TOSession.CreateDeliveryServiceRequest(dsr)
+		_, _, _, err := TOSession.CreateDeliveryServiceRequest(dsr, nil)
 		if err == nil {
 			t.Error("expected: validation error, actual: nil")
 		}
@@ -141,24 +181,39 @@ func TestDeliveryServiceRequestRules(t *testing.T) {
 func TestDeliveryServiceRequestTypeFields(t *testing.T) {
 	WithObjs(t, []TCObj{CDNs, Types, Tenants, Parameters}, func() {
 		dsr := testData.DeliveryServiceRequests[dsrBadTenant]
+		resetDS(dsr.DeliveryService)
+		if dsr.DeliveryService == nil || dsr.DeliveryService.XMLID == nil {
+			t.Fatalf("the %dth DSR in the test data had no DeliveryService - or that DeliveryService had no XMLID", dsrBadTenant)
+		}
 
-		alerts, _, err := TOSession.CreateDeliveryServiceRequest(dsr)
+		_, alerts, _, err := TOSession.CreateDeliveryServiceRequest(dsr, nil)
 		if err != nil {
 			t.Errorf("Error occurred %v", err)
 		}
 
-		expected := []string{
-			"deliveryservice_request was created.",
-			//"'xmlId' the length must be between 1 and 48",
+		found := false
+		for _, alert := range alerts.Alerts {
+			if alert.Level == tc.ErrorLevel.String() {
+				t.Errorf("Expected only succuss-level alerts creating a DSR, got error-level alert: %s", alert.Text)
+			} else if alert.Level == tc.SuccessLevel.String() {
+				t.Logf("Got expected alert creating a DSR: %s", alert.Text)
+				found = true
+			}
 		}
 
-		utils.Compare(t, expected, alerts.ToStrings())
+		if !found {
+			t.Errorf("Expected a success-level alert creating a DSR, got none")
+		}
 
-		dsrs, _, err := TOSession.GetDeliveryServiceRequestByXMLID(dsr.DeliveryService.XMLID)
+		dsrs, _, _, err := TOSession.GetDeliveryServiceRequestsByXMLID(*dsr.DeliveryService.XMLID, nil)
 		if len(dsrs) != 1 {
-			t.Errorf("expected 1 deliveryservice_request with XMLID %s;  got %d", dsr.DeliveryService.XMLID, len(dsrs))
+			t.Fatalf("expected 1 deliveryservice_request with XMLID %s;  got %d", *dsr.DeliveryService.XMLID, len(dsrs))
 		}
-		alert, _, err := TOSession.DeleteDeliveryServiceRequestByID(dsrs[0].ID)
+		if dsrs[0].ID == nil {
+			t.Fatalf("got a DSR with a null ID by XMLID '%s'", *dsr.DeliveryService.XMLID)
+		}
+
+		alert, _, err := TOSession.DeleteDeliveryServiceRequest(*dsrs[0].ID)
 		if err != nil {
 			t.Errorf("cannot DELETE DeliveryServiceRequest by id: %d - %v - %v", dsrs[0].ID, err, alert)
 		}
@@ -169,19 +224,16 @@ func TestDeliveryServiceRequestBad(t *testing.T) {
 	WithObjs(t, []TCObj{CDNs, Types, Parameters, Tenants}, func() {
 		// try to create non-draft/submitted
 		src := testData.DeliveryServiceRequests[dsrDraft]
+		resetDS(src.DeliveryService)
 		s, err := tc.RequestStatusFromString("pending")
 		if err != nil {
 			t.Errorf(`unable to create Status from string "pending"`)
 		}
 		src.Status = s
 
-		_, _, err = TOSession.CreateDeliveryServiceRequest(src)
+		_, _, _, err = TOSession.CreateDeliveryServiceRequest(src, nil)
 		if err == nil {
 			t.Fatal("expected: validation error, actual: nil")
-		}
-		expected := `'status' invalid transition from draft to pending`
-		if !strings.Contains(err.Error(), expected) {
-			t.Fatalf("expected: error message to contain %s, actual: %s", expected, err.Error())
 		}
 	})
 }
@@ -190,7 +242,7 @@ func TestDeliveryServiceRequestBad(t *testing.T) {
 func TestDeliveryServiceRequestWorkflow(t *testing.T) {
 	WithObjs(t, []TCObj{CDNs, Types, Parameters, Tenants}, func() {
 		// test empty request table
-		dsrs, _, err := TOSession.GetDeliveryServiceRequests()
+		dsrs, _, _, err := TOSession.GetDeliveryServiceRequests(nil)
 		if err != nil {
 			t.Errorf("Error getting empty list of DeliveryServiceRequests %v++", err)
 		}
@@ -203,28 +255,50 @@ func TestDeliveryServiceRequestWorkflow(t *testing.T) {
 
 		// Create a draft request
 		src := testData.DeliveryServiceRequests[dsrDraft]
+		resetDS(src.DeliveryService)
 
-		alerts, _, err := TOSession.CreateDeliveryServiceRequest(src)
+		_, alerts, _, err := TOSession.CreateDeliveryServiceRequest(src, nil)
 		if err != nil {
 			t.Errorf("Error creating DeliveryServiceRequest %v", err)
 		}
 
-		expected := []string{`deliveryservice_request was created.`}
-		utils.Compare(t, expected, alerts.ToStrings())
+		found := false
+		for _, alert := range alerts.Alerts {
+			if alert.Level == tc.ErrorLevel.String() {
+				t.Errorf("Expected only succuss-level alerts creating a DSR, got error-level alert: %s", alert.Text)
+			} else if alert.Level == tc.SuccessLevel.String() {
+				t.Logf("Got expected alert creating a DSR: %s", alert.Text)
+				found = true
+			}
+		}
+
+		if !found {
+			t.Errorf("Expected a success-level alert creating a DSR, got none")
+		}
 
 		// Create a duplicate request -- should fail because xmlId is the same
-		alerts, _, err = TOSession.CreateDeliveryServiceRequest(src)
+		_, alerts, _, err = TOSession.CreateDeliveryServiceRequest(src, nil)
 		if err == nil {
 			t.Fatal("expected: validation error, actual: nil")
 		}
 
-		expectedStr := `An active request exists for delivery service 'test-transitions'`
-		if !strings.Contains(err.Error(), expectedStr) {
-			t.Errorf("expected: error message containing %s, actual: %v", expectedStr, err)
-		}
-		utils.Compare(t, expected, alerts.ToStrings())
+		// TODO: the client needs to be fixed to return alerts on error. But
+		// that's gotten harder to do "semantically" now that most of the logic
+		// is shared between all API versions.
+		// found = false
+		// for _, alert := range alerts.Alerts {
+		// 	if alert.Level == tc.SuccessLevel.String() {
+		// 		t.Errorf("Expected only error-level alerts creating a duplicate DSR, got success-level alert: %s", alert.Text)
+		// 	} else if alert.Level == tc.ErrorLevel.String() {
+		// 		t.Logf("Got expected alert creating a duplicate DSR: %s", alert.Text)
+		// 		found = true
+		// 	}
+		// }
+		// if !found {
+		// 	t.Errorf("Expected an error-level alert creating a duplicate DSR, got none")
+		// }
 
-		dsrs, _, err = TOSession.GetDeliveryServiceRequestByXMLID(`test-transitions`)
+		dsrs, _, _, err = TOSession.GetDeliveryServiceRequestsByXMLID(`test-transitions`, nil)
 		if len(dsrs) != 1 {
 			t.Errorf("Expected 1 deliveryServiceRequest -- got %d", len(dsrs))
 			if len(dsrs) == 0 {
@@ -234,37 +308,46 @@ func TestDeliveryServiceRequestWorkflow(t *testing.T) {
 
 		alerts, dsr := updateDeliveryServiceRequestStatus(t, dsrs[0], "submitted", nil)
 
-		expected = []string{
-			"deliveryservice_request was updated.",
+		found = false
+		for _, alert := range alerts.Alerts {
+			if alert.Level == tc.ErrorLevel.String() {
+				t.Errorf("Expected only succuss-level alerts updating a DSR, got error-level alert: %s", alert.Text)
+			} else if alert.Level == tc.SuccessLevel.String() {
+				t.Logf("Got expected alert updating a DSR: %s", alert.Text)
+				found = true
+			}
+		}
+		if !found {
+			t.Errorf("Expected a success-level alert updating a DSR, got none")
 		}
 
-		utils.Compare(t, expected, alerts.ToStrings())
 		if dsr.Status != tc.RequestStatus("submitted") {
 			t.Errorf("expected status=submitted,  got %s", string(dsr.Status))
 		}
 	})
 }
 
-func updateDeliveryServiceRequestStatus(t *testing.T, dsr tc.DeliveryServiceRequest, newstate string, header http.Header) (tc.Alerts, tc.DeliveryServiceRequest) {
+func updateDeliveryServiceRequestStatus(t *testing.T, dsr tc.DeliveryServiceRequestV40, newstate string, header http.Header) (tc.Alerts, tc.DeliveryServiceRequestV40) {
 	ID := dsr.ID
+	if ID == nil {
+		t.Error("updateDeliveryServiceRequestStatus called with a DSR that has a nil ID")
+		return tc.Alerts{}, tc.DeliveryServiceRequestV40{}
+	}
 	dsr.Status = tc.RequestStatus("submitted")
 
-	alerts, _, err := TOSession.UpdateDeliveryServiceRequestByIDWithHdr(ID, dsr, header)
+	_, alerts, _, err := TOSession.UpdateDeliveryServiceRequest(*ID, dsr, header)
 	if err != nil {
 		t.Errorf("Error updating deliveryservice_request: %v", err)
 		return alerts, dsr
 	}
 
-	d, _, err := TOSession.GetDeliveryServiceRequestByID(ID)
+	d, _, _, err := TOSession.GetDeliveryServiceRequest(*ID, nil)
 	if err != nil {
 		t.Errorf("Error updating deliveryservice_request %d: %v", ID, err)
 		return alerts, dsr
 	}
 
-	if len(d) != 1 {
-		t.Errorf("Expected 1 deliveryservice_request, got %d", len(d))
-	}
-	return alerts, d[0]
+	return alerts, d
 }
 
 func GetTestDeliveryServiceRequestsIMS(t *testing.T) {
@@ -274,7 +357,12 @@ func GetTestDeliveryServiceRequestsIMS(t *testing.T) {
 	time := futureTime.Format(time.RFC1123)
 	header.Set(rfc.IfModifiedSince, time)
 	dsr := testData.DeliveryServiceRequests[dsrGood]
-	_, reqInf, err := TOSession.GetDeliveryServiceRequestByXMLIDWithHdr(dsr.DeliveryService.XMLID, header)
+	resetDS(dsr.DeliveryService)
+	if dsr.DeliveryService == nil || dsr.DeliveryService.XMLID == nil {
+		t.Fatalf("the %dth DSR in the test data had no DeliveryService - or that DeliveryService had no XMLID", dsrGood)
+	}
+
+	_, _, reqInf, err := TOSession.GetDeliveryServiceRequestsByXMLID(*dsr.DeliveryService.XMLID, header)
 	if err != nil {
 		t.Fatalf("Expected no error, but got %v", err.Error())
 	}
@@ -285,7 +373,11 @@ func GetTestDeliveryServiceRequestsIMS(t *testing.T) {
 
 func GetTestDeliveryServiceRequests(t *testing.T) {
 	dsr := testData.DeliveryServiceRequests[dsrGood]
-	resp, _, err := TOSession.GetDeliveryServiceRequestByXMLID(dsr.DeliveryService.XMLID)
+	resetDS(dsr.DeliveryService)
+	if dsr.DeliveryService == nil || dsr.DeliveryService.XMLID == nil {
+		t.Fatalf("the %dth DSR in the test data had no DeliveryService - or that DeliveryService had no XMLID", dsrGood)
+	}
+	resp, _, _, err := TOSession.GetDeliveryServiceRequestsByXMLID(*dsr.DeliveryService.XMLID, nil)
 	if err != nil {
 		t.Errorf("cannot GET DeliveryServiceRequest by XMLID: %v - %v", err, resp)
 	}
@@ -295,7 +387,12 @@ func UpdateTestDeliveryServiceRequests(t *testing.T) {
 
 	// Retrieve the DeliveryServiceRequest by name so we can get the id for the Update
 	dsr := testData.DeliveryServiceRequests[dsrGood]
-	resp, _, err := TOSession.GetDeliveryServiceRequestByXMLID(dsr.DeliveryService.XMLID)
+	resetDS(dsr.DeliveryService)
+	if dsr.DeliveryService == nil || dsr.DeliveryService.XMLID == nil {
+		t.Fatalf("the %dth DSR in the test data had no DeliveryService - or that DeliveryService had no XMLID", dsrGood)
+	}
+
+	resp, _, _, err := TOSession.GetDeliveryServiceRequestsByXMLID(*dsr.DeliveryService.XMLID, nil)
 	if err != nil {
 		t.Errorf("cannot GET DeliveryServiceRequest by name: %v - %v", dsr.DeliveryService.XMLID, err)
 	}
@@ -303,10 +400,14 @@ func UpdateTestDeliveryServiceRequests(t *testing.T) {
 		t.Fatal("Length of GET DeliveryServiceRequest is 0")
 	}
 	respDSR := resp[0]
+	if respDSR.ID == nil {
+		t.Fatalf("got a DSR by XMLID '%s' with a null or undefined ID", *dsr.DeliveryService.XMLID)
+	}
 	expDisplayName := "new display name"
-	respDSR.DeliveryService.DisplayName = expDisplayName
+	respDSR.DeliveryService.DisplayName = &expDisplayName
 	var alert tc.Alerts
-	alert, _, err = TOSession.UpdateDeliveryServiceRequestByID(respDSR.ID, respDSR)
+	id := *respDSR.ID
+	_, alert, _, err = TOSession.UpdateDeliveryServiceRequest(id, respDSR, nil)
 	t.Log("Response: ", alert)
 	if err != nil {
 		t.Errorf("cannot UPDATE DeliveryServiceRequest by id: %v - %v", err, alert)
@@ -314,39 +415,50 @@ func UpdateTestDeliveryServiceRequests(t *testing.T) {
 	}
 
 	// Retrieve the DeliveryServiceRequest to check DeliveryServiceRequest name got updated
-	resp, _, err = TOSession.GetDeliveryServiceRequestByID(respDSR.ID)
+	respDSR, _, _, err = TOSession.GetDeliveryServiceRequest(id, nil)
 	if err != nil {
-		t.Errorf("cannot GET DeliveryServiceRequest by name: %v - %v", respDSR.ID, err)
-	} else {
-		respDSR = resp[0]
-		if respDSR.DeliveryService.DisplayName != expDisplayName {
-			t.Errorf("results do not match actual: %s, expected: %s", respDSR.DeliveryService.DisplayName, expDisplayName)
-		}
+		t.Fatalf("cannot GET DeliveryServiceRequest by ID: %v - %v", id, err)
 	}
-
+	if respDSR.DeliveryService == nil || respDSR.DeliveryService.DisplayName == nil {
+		t.Fatalf("Got DSR by ID '%d' that had no DeliveryService - or said DeliveryService had no DisplayName", *respDSR.ID)
+	}
+	if *respDSR.DeliveryService.DisplayName != expDisplayName {
+		t.Errorf("results do not match actual: %s, expected: %s", *respDSR.DeliveryService.DisplayName, expDisplayName)
+	}
 }
 
 func DeleteTestDeliveryServiceRequests(t *testing.T) {
 
 	// Retrieve the DeliveryServiceRequest by name so we can get the id for the Update
 	dsr := testData.DeliveryServiceRequests[dsrGood]
-	resp, _, err := TOSession.GetDeliveryServiceRequestByXMLID(dsr.DeliveryService.XMLID)
+	resetDS(dsr.DeliveryService)
+	if dsr.DeliveryService == nil || dsr.DeliveryService.XMLID == nil {
+		t.Fatalf("the %dth DSR in the test data had no DeliveryService - or that DeliveryService had no XMLID", dsrGood)
+	}
+
+	resp, _, _, err := TOSession.GetDeliveryServiceRequestsByXMLID(*dsr.DeliveryService.XMLID, nil)
 	if err != nil {
-		t.Errorf("cannot GET DeliveryServiceRequest by id: %v - %v", dsr.DeliveryService.XMLID, err)
+		t.Fatalf("cannot GET DeliveryServiceRequest by id: %v - %v", dsr.DeliveryService.XMLID, err)
+	}
+	if len(resp) < 1 {
+		t.Fatal("expected at least one Delivery Service Request, got none")
 	}
 	respDSR := resp[0]
-	alert, _, err := TOSession.DeleteDeliveryServiceRequestByID(respDSR.ID)
+	if respDSR.ID == nil {
+		t.Fatalf("Got a DSR by XMLID '%s' that had no ID", *dsr.DeliveryService.XMLID)
+	}
+	alert, _, err := TOSession.DeleteDeliveryServiceRequest(*respDSR.ID)
 	t.Log("Response: ", alert)
 	if err != nil {
 		t.Errorf("cannot DELETE DeliveryServiceRequest by id: %d - %v - %v", respDSR.ID, err, alert)
 	}
 
 	// Retrieve the DeliveryServiceRequest to see if it got deleted
-	dsrs, _, err := TOSession.GetDeliveryServiceRequestByXMLID(dsr.DeliveryService.XMLID)
+	dsrs, _, _, err := TOSession.GetDeliveryServiceRequestsByXMLID(*dsr.DeliveryService.XMLID, nil)
 	if err != nil {
 		t.Errorf("error deleting DeliveryServiceRequest name: %s", err.Error())
 	}
 	if len(dsrs) > 0 {
-		t.Errorf("expected DeliveryServiceRequest XMLID: %s to be deleted", dsr.DeliveryService.XMLID)
+		t.Errorf("expected DeliveryServiceRequest XMLID: %s to be deleted", *dsr.DeliveryService.XMLID)
 	}
 }